Struct marksman_escape::Unescape

#[must_use = "iterator adaptors are lazy and do nothing unless consumed"]
pub struct Unescape<I: Iterator<Item=u8>> {
    // some fields omitted
}

Unescape a HTML-encoded stream of bytes.

The HTML5 named character references (&amp;), decimal character references (&#123;) and hexadecimal character references (&#x1BA;) are supported.

The implementation works with bytes interpreting them to be ASCII, which means that any ASCII-compatible encoding, including UTF-8, is supported.

Methods

impl<I: Iterator<Item=u8>> Unescape<I>

fn new(i: I) -> Unescape<I>

Create an iterator adaptor which will unescape all the character references found in the internal iterator.

Usage

use marksman_escape::Unescape;
let string = "&lt;hello&gt;&amp;world&#60;/hello&#x3e;";
let unescaped = String::from_utf8(Unescape::new(string.bytes()).collect()).unwrap();
assert_eq!("<hello>&world</hello>", &*unescaped);

impl<I: Iterator<Item=u8>> Unescape<I>

fn new(i: I) -> Unescape<I>

Create an iterator adaptor which will unescape all the character references found in the internal iterator.

Usage

use marksman_escape::Unescape;
let string = "&lt;hello&gt;&amp;world&#60;/hello&#x3e;";
let unescaped = String::from_utf8(Unescape::new(string.bytes()).collect()).unwrap();
assert_eq!("<hello>&world</hello>", &*unescaped);